Move rpcconsumerlogs into common module
